backend/requirement.txt
fastapi
uvicorn[standard]
sqlalchemy
bcrypt==4.0.1
python-dotenv
httpx
pydantic
python-multipart
Pillow